Move promise queue
