+++ packages/spatial/src/common/classes/PromiseQueue.ts
